Fix minor .ma bug
Removed unnecessary "i -= 1" statements from the code that
processes the .ma directive. As each section of code deletes
the .ma statement, then uses continue, the loop variable i is
never incremented when the .ma statement is processed, and so
the variable does not need to be decremented.

With the decrement in place, the statement before the .ma is scanned
multiple times. Or, for a .ma happening first in the file, the last
statement in the file is processed unnecessarily when looking at
after i is decremented from 0 to -1 after that first statement is
deleted and we then look at self.wb[-1].