Removed erroneous "i -= 1" while merging page numbers down
Just after a del self.wb[i] when finding two .pn commands in a row
ppgen had
  i -= 1  counteract i += 1 below
  continue
The decrement is unneeded because with the "continue" it never reaches
the increment operation. In practice, the decrement can cause the page
number to actually move up a line into text rather than down as intended.