Footnote handling inconsistencies in text and HTML
(1) HTML footnote handling incremented fncr when processing an explicit
footnote number in a reference or footnote definition, but text footnote
processing did not do that.

(2) However, the code in HTML processing did not allow for the possibility
of more than one explicit footnote reference in a source line, and only
incremented fncr for the first one in the line.

Both have been fixed, by correcting the code for (2) and copying that code
and the code for (1) into the text portion of ppgen.

Note: The code (still) does not support intermixing explicit (numbers) and
implicit (#) references within the same source line. A single source line can
have either explicit or implicit references, but not both.