3.43d Fix erroneous sc warning within .nf block
Normally ppgen will warn the PPer if it sees an all lower-case string of
text within sc /sc tags, as we never want all lower-case small-cap
strings. That case should be prepared by the formatters and the PPer
as all upper-case within the tags, instead.

However, for text within .nf blocks ppgen outputs each line in a div, and
so it needs to replicate any inline tags that span lines. This can result
in a line that is both all lower-case and within the sc tags, although the
complete string as prepared by the formatters and PPer was in fact
mixed-case. So, within a .nf block, ppgen will suppress that warning
message.